Move test into existing module
